Refactor CI workflow for MinGW; remove unnecessary make program configuration and streamline Conan install command Enhance file_header.hpp; include <array> for improved functionality Fix type casting in file_header.cpp; ensure correct offset handling during data copy Update progress.cpp; cast max_progress to float for proper progress bar functionality

+++ src/utils/progress.cpp
@@ -93,7 +93,7 @@ BlockProgressBar::BlockProgressBar(const std::string& prefix, size_t max_progres
         option::ForegroundColor{Color::cyan},
         option::PrefixText{prefix},
         option::ShowPercentage{true},
-        option::MaxProgress{max_progress}
+        option::MaxProgress{static_cast<float>(max_progress)}
     );
 }
